backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
django-cors-headers==4.3.1
pillow==10.2.0